requirements.txt
torch
transformers
peft
datasets
accelerate
python-dotenv
pyyaml
huggingface_hub 